Put logic inside controller
